fragments + hero actions
