+++ heroes.client.ngx-app/src/app/shared/index.ts
@@ -7,6 +7,7 @@ export * from "./state/hero-category.state";
 
 export * from "./state/hero.model";
 export * from "./state/hero.service";
+export * from "./state/hero.action";
 export * from "./state/hero.state";
 
 export * from "./app-info.service";
