+++ heroes.client.ngx-app/src/app/shared/state/hero-category.action.ts
@@ -1,0 +1,9 @@
+
+export namespace HeroCategoryActions {
+
+
+	export class Load {
+		static readonly type = "[Hero] Load";
+	}
+
+}
