+++ heroes.client.ngx-app/src/app/shared/state/hero.state.ts
@@ -8,49 +8,7 @@ import { Dictionary } from "@ssv/core";
 import { arrayToObject } from "../utils";
 import { Hero } from "./hero.model";
 import { HeroService } from "./hero.service";
-
-export namespace HeroActions {
-
-	export class Add {
-		static readonly type = "[Hero] Add";
-
-		constructor(
-			public payload: Hero
-		) {
-		}
-	}
-
-	export class Update {
-		static readonly type = "[Hero] Update";
-
-		constructor(
-			public payload: Hero
-		) {
-		}
-	}
-
-	export class UpdateAll {
-		static readonly type = "[Hero] UpdateAll";
-
-		constructor(
-			public payload: Hero[]
-		) {
-		}
-	}
-
-	export class Load {
-		static readonly type = "[Hero] Load";
-	}
-
-	export class Select {
-		static readonly type = "[Hero] Select";
-		constructor(
-			public key: string
-		) {
-		}
-	}
-
-}
+import { HeroActions } from "./hero.action";
 
 export interface HeroStateModel {
 	entities: Dictionary<Hero>;
